RBF-FD-Prototype
Overview: This is a prototype PDE solver. Supported time integration schemes are either RK3 or RK4.
The model is designed to easily integate new systems of equations by creating a new rhs
function f* in rhs.c.

Language. The code is written in C, but be advised it needs a lot of cleaning up.

Precision. The code supports easy toggling between float or double execution.
To change precision, go to floating_types.h and toggle the fType typedef.

Compilers. The code has only been tested on the Clang 11 compiler.

Parallelism. It is serial - all MPI, OpenMP, OpenCL and OpenACC has been stripped out.